Split integration tests
