+++ .github/workflows/integration_test.yaml
@@ -10,11 +10,12 @@ jobs:
     with:
       load-test-enabled: false
       load-test-run-args: "-e LOAD_TEST_HOST=localhost"
+      modules: '["test_actions.py", "test_charm.py", "test_saml.py"]'
+      trivy-fs-enabled: true
+      trivy-image-config: "trivy.yaml"
       zap-before-command: "curl -H \"Host: indico.local\" http://localhost/bootstrap --data-raw 'csrf_token=00000000-0000-0000-0000-000000000000&first_name=admin&last_name=admin&email=admin%40admin.com&username=admin&password=lunarlobster&confirm_password=lunarlobster&affiliation=Canonical'"
       zap-enabled: true
       zap-cmd-options: '-T 60 -z "-addoninstall jython" --hook "/zap/wrk/tests/zap/hook.py"'
       zap-target: localhost
       zap-target-port: 80
       zap-rules-file-name: "zap_rules.tsv"
-      trivy-fs-enabled: true
-      trivy-image-config: "trivy.yaml"
