+++ tests/integration/test_charm.py
@@ -88,40 +88,6 @@ async def test_health_checks(app: Application):
         assert stdout.count("0/3") == container_checks[1]
 
 
-@pytest.mark.abort_on_fail
-@pytest_asyncio.fixture(scope="module")
-async def add_admin(app: Application):
-    """
-    arrange: given charm in its initial state
-    act: run the add-admin action
-    assert: check the output in the action result
-    """
-    assert hasattr(app, "units")
-
-    assert app.units[0]
-
-    email = ADMIN_USER_EMAIL
-    email_fail = ADMIN_USER_EMAIL_FAIL
-    # This is a test password
-    password = "somepassword"  # nosec
-
-    # Application actually does have units
-    action: juju.action.Action = await app.units[0].run_action(  # type: ignore
-        "add-admin", email=email, password=password
-    )
-    await action.wait()
-    assert action.status == "completed"
-    assert action.results["user"] == email
-    assert f'Admin with email "{email}" correctly created' in action.results["output"]
-    action2: juju.action.Action = await app.units[0].run_action(  # type: ignore
-        "add-admin", email=email_fail, password=password
-    )
-    await action2.wait()
-    assert action2.status == "completed"
-    assert action2.results["user"] == email_fail
-    assert f'Admin with email "{email_fail}" correctly created' in action2.results["output"]
-
-
 @pytest.mark.asyncio
 @pytest.mark.abort_on_fail
 async def test_prom_exporters_are_up(app: Application):
@@ -148,140 +114,3 @@ async def test_prom_exporters_are_up(app: Application):
         stdout = result["results"].get("Stdout")
         stderr = result["results"].get("Stderr")
         assert code == "0", f"{cmd} failed ({code}): {stderr or stdout}"
-
-
-@pytest.mark.asyncio
-@pytest.mark.abort_on_fail
-@pytest.mark.usefixtures("add_admin")
-async def test_anonymize_user(app: Application):
-    """
-    arrange: admin user created
-    act: run the anonymize-user action
-    assert: check the output in the action result
-    """
-    # Application actually does have units
-    action_anonymize: juju.action.Action = await app.units[0].run_action(  # type: ignore
-        "anonymize-user", email=ADMIN_USER_EMAIL
-    )
-    await action_anonymize.wait()
-    assert action_anonymize.status == "completed"
-    assert action_anonymize.results["user"] == ADMIN_USER_EMAIL
-    expected_words = [ADMIN_USER_EMAIL, "correctly anonymized"]
-    assert all(word in action_anonymize.results["output"] for word in expected_words)
-
-
-@pytest.mark.asyncio
-@pytest.mark.abort_on_fail
-@pytest.mark.usefixtures("add_admin")
-async def test_anonymize_user_fail(app: Application):
-    """
-    arrange: admin user created
-    act: run the anonymize-user action
-    assert: check the output in the action result
-    """
-    # Application actually does have units
-    action_anonymize: juju.action.Action = await app.units[0].run_action(  # type: ignore
-        "anonymize-user", email=f",{ADMIN_USER_EMAIL_FAIL}"
-    )
-    await action_anonymize.wait()
-    assert action_anonymize.status == "failed"
-    assert action_anonymize.results["user"] == f",{ADMIN_USER_EMAIL_FAIL}"
-    expected_words = [ADMIN_USER_EMAIL_FAIL, "correctly anonymized", "Failed to anonymize user"]
-    assert all(word in action_anonymize.results["output"] for word in expected_words)
-
-
-@pytest.mark.asyncio
-@pytest.mark.abort_on_fail
-async def test_saml_auth(
-    ops_test: OpsTest,
-    app: Application,
-    saml_email: str,
-    saml_password: str,
-    requests_timeout: float,
-):
-    """
-    arrange: given charm in its initial state
-    act: configure a SAML target url and fire SAML authentication
-    assert: The SAML authentication process is executed successfully.
-    """
-    # The linter does not recognize set_config as a method, so this errors must be ignored.
-    await app.set_config(  # type: ignore[attr-defined] # pylint: disable=W0106
-        {
-            "site_url": "https://events.staging.canonical.com",
-            "saml_target_url": STAGING_UBUNTU_SAML_URL,
-        }
-    )
-    # The linter does not recognize wait_for_idle as a method,
-    # since ops_test has a model as Optional, so this error must be ignored.
-    await ops_test.model.wait_for_idle(status="active")  # type: ignore[union-attr]
-    urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-
-    host = "events.staging.canonical.com"
-    original_getaddrinfo = socket.getaddrinfo
-
-    def patched_getaddrinfo(*args):
-        """Get address info forcing localhost as the IP.
-
-        Args:
-            args: Arguments from the getaddrinfo original method.
-
-        Returns:
-            Address information with localhost as the patched IP.
-        """
-        if args[0] == host:
-            return original_getaddrinfo("127.0.0.1", *args[1:])
-        return original_getaddrinfo(*args)
-
-    with patch.multiple(socket, getaddrinfo=patched_getaddrinfo), requests.session() as session:
-        session.get(f"https://{host}", verify=False)
-        login_page = session.get(
-            f"https://{host}/login",
-            verify=False,
-            timeout=requests_timeout,
-        )
-        csrf_token_matches = re.findall(
-            "<input type='hidden' name='csrfmiddlewaretoken' value='([^']+)' />", login_page.text
-        )
-        assert csrf_token_matches, login_page.text
-        saml_callback = session.post(
-            "https://login.staging.ubuntu.com/+login",
-            data={
-                "csrfmiddlewaretoken": csrf_token_matches[0],
-                "email": saml_email,
-                "user-intentions": "login",
-                "password": saml_password,
-                "next": "/saml/process",
-                "continue": "",
-                "openid.usernamesecret": "",
-                "RelayState": "indico.local",
-            },
-            headers={"Referer": login_page.url},
-            timeout=requests_timeout,
-        )
-        saml_response_matches = re.findall(
-            '<input type="hidden" name="SAMLResponse" value="([^"]+)" />', saml_callback.text
-        )
-        assert len(saml_response_matches), saml_callback.text
-        session.post(
-            f"https://{host}/multipass/saml/ubuntu/acs",
-            data={
-                "RelayState": "None",
-                "SAMLResponse": saml_response_matches[0],
-                "openid.usernamesecret": "",
-            },
-            verify=False,
-            timeout=requests_timeout,
-        )
-        session.post(
-            f"https://{host}/multipass/saml/ubuntu/acs",
-            data={"SAMLResponse": saml_response_matches[0], "SameSite": "1"},
-            verify=False,
-            timeout=requests_timeout,
-        )
-
-        dashboard_page = session.get(
-            f"https://{host}/register/ubuntu",
-            verify=False,
-            timeout=requests_timeout,
-        )
-        assert dashboard_page.status_code == 200
